+++ test/unit/gonzales/adapter_test.rb
@@ -21,31 +21,19 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'factory_girl'
-module Gonzales
-  # = Gonzales::TestHelper
-  #
-  # Helper to be used when writing tests when you use Gonzales
-  module TestHelper
-    # returns a created record to be used in tests
-    #
-    # speedy will return the record from the database if it alreday exists in the test database.
-    # If not it will call FactoryGirl to instantiate the record.
-    #
-    # === Arguments
-    #
-    #   * factory_name - Name of the factory or the alias specified in the gonzales file
-    #   * options - will be passed to FactoryGirl if the factory does not exist in the test database.
-    #     Note!! options will not be passed to Gonzales. 
-    #
-    # === Example
-    #
-    #   setup do
-    #     @person = speedy :person
-    #   end
-    #
-    def speedy(factory_name, *options)
-      Collection.entity(factory_name) || Adapter.create(factory_name, *options)
+require 'test_helper'
+
+class Gonzales::AdapterTest < ActiveSupport::TestCase
+  context 'creation' do
+    should 'default to unregistered' do
+      Gonzales::Adapter.class_variable_set(:@@adapter, nil)
+      Gonzales::Adapter::Unregistered.expects(:create).with(:cartoon, :elmer)
+      Gonzales::Adapter.create(:cartoon, :elmer)
+    end
+    should 'use the adapter specified' do
+      Gonzales::Adapter.use :registered
+      Gonzales::Adapter::Registered.expects(:create).with(:cartoon, :elmer)
+      Gonzales::Adapter.create(:cartoon, :elmer)
     end
   end
 end