+++ release_notes.md
@@ -2,9 +2,11 @@
 
 * Initial gem release
 
-## 0.5.1
+## 0.6.0
 
 ### New features
 
 * Added configuration - disable_preload for use with integration tests
+* Added configuration - adapter to control caching of factory references. Relevant for integration tests
+* Changed the way the factory references are handled when they are loaded